+++ cmd/hidden_lake/service/internal/handler/handler_test.go
@@ -185,9 +185,7 @@ func testRunNewNode(dbPath, addr string) (anonymity.INode, context.Context, cont
 func testNewNode(dbPath, addr string) anonymity.INode {
 	db, err := database.NewKVDatabase(
 		database.NewSettings(&database.SSettings{
-			FPath:     dbPath,
-			FWorkSize: testutils.TCWorkSize,
-			FPassword: "CIPHER",
+			FPath: dbPath,
 		}),
 	)
 	if err != nil {