+++ pkg/crypto/keybuilder/types.go
@@ -1,5 +1,0 @@
-package keybuilder
-
-type IKeyBuilder interface {
-	Build(string, uint64) []byte
-}